api: Fix missing endpoint in OpenAPI (#7214)
