+++ hoprd/rest-api/src/middleware/preconditions.rs
