+++ hoprd/rest-api/src/middleware/prometheus.rs
